requirements.txt
pvporcupine
sounddevice
numpy
SpeechRecognition
pyttsx3
PyYAML
requests
wikipedia
pycaw
comtypes
screen-brightness-control
